% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_weights.R
\name{create_pmatrix_from_terms}
\alias{create_pmatrix_from_terms}
\title{Create Matrix from Select Terms}
\usage{
create_pmatrix_from_terms(xx, terms)
}
\arguments{
\item{xx}{raw matrix with colnames with natural indexes}

\item{terms}{selected terms. If exist interact terms like \code{i*j}, then
the i-th column would multiply j-th column, and result in a new column.}
}
\value{
a matrix
}
\description{
Create Matrix from Select Terms
}
