% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{find.index}
\alias{find.index}
\title{match index}
\usage{
find.index(df1, df2, type = "reg")
}
\arguments{
\item{df1}{data.frame 1}

\item{df2}{data.frame 2}

\item{type}{type}
}
\value{
matched index
}
\description{
match index
}
