% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{format_twas_dat}
\alias{format_twas_dat}
\title{Data preparation}
\usage{
format_twas_dat(dat, E, E.info)
}
\arguments{
\item{dat}{plink bfile}

\item{E}{original gene expression matrix (list), the length of the list is the number of tissues,  each element is a sample*gene matrix, rownames are sample id, colnames are gene id}

\item{E.info}{data.frame with columns: chr, start, end, pos}
}
\value{
standard mtwas data format
}
\description{
Data preparation
}
