% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute.E}
\alias{impute.E}
\title{Impute the list of expression matrix `E`}
\usage{
impute.E(x)
}
\arguments{
\item{x}{twas data list}
}
\value{
imputed twas data list and the imputation MSE error
}
\description{
Impute the list of expression matrix `E`
}
