% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twas.R
\name{run_mtwas_easy}
\alias{run_mtwas_easy}
\title{Run association test with downloaded trained data}
\usage{
run_mtwas_easy(stats, twas_bim, twas_eqtl)
}
\arguments{
\item{stats}{summary statistics for a phenotype. It can be a filename to be read or a data.frame, should include columns rsid, a1, a2, chr, z}

\item{twas_bim}{pre-downloaded file}

\item{twas_eqtl}{pre-downloaded file}
}
\value{
mtwas summary statistics
}
\description{
Run association test with downloaded trained data
}
