% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_weights.R
\name{select.ct.eQTL}
\alias{select.ct.eQTL}
\title{Detect the cross-tissue eQTLs}
\usage{
select.ct.eQTL(
  x,
  nne = -1,
  gam = 1,
  nF0 = 5,
  verbose = T,
  allow_empty = T,
  ncores = 1L
)
}
\arguments{
\item{x}{TWAS data, which is a list of three elements: \describe{
 \item{E}{list of cell types, each element is a matrix of size n_sample x n_gene}
 \item{dat}{a list with three named elements. \describe{
   \item{bed}{a matrix of size n_sample x n_snp}
   \item{bim}{a data frame with n_snp rows, each row is the info for the snp}}
 }
 \item{pos}{a list of genes, each element is a vector containing the SNP index that
 locate in the gene, and it corresponds to each row of \code{E.info}.
 It can be derived from \code{E.info} and \code{dat}.}
}}

\item{nne}{number of genes to be considered. Default is -1, which means to use all genes.}

\item{gam}{the tuning parameter \eqn{\gamma} in the EBIC criterion. Default is 1.}

\item{nF0}{the smallest number of main effects. Default is 5.}

\item{verbose}{whether to print running message. Default is \code{TRUE}.}

\item{allow_empty}{whether to allow empty set. Default is \code{TRUE}.}

\item{ncores}{if using parallel computing, the number of cores to be used. Default is 1.}
}
\value{
a list of size n_gene, each element is the shared eQTL for the gene.
}
\description{
Detect the cross-tissue eQTLs
}
