% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_weights.R
\name{soda_model_imp_share}
\alias{soda_model_imp_share}
\title{Select cross-tissue via SODA}
\usage{
soda_model_imp_share(
  j,
  gam,
  E,
  dat,
  pos,
  nF0 = 5,
  verbose = TRUE,
  allow_empty = TRUE
)
}
\arguments{
\item{j}{the index of gene}

\item{gam}{TODO}

\item{E}{list of cell types, each element is a matrix of size n_sample x n_gene}

\item{dat}{a list with three named elements. \describe{
\item{bed}{a matrix of size n_sample x n_snp}
\item{bim}{a data frame with n_snp rows, each row is the info for the snp}
\item{fam}{a data frame with n_sample rows, each row is the info for the sample}}}

\item{pos}{a list of genes, each element is a vector containing the SNP index that
locate in the gene, and it corresponds to each row of \code{E.info}.
It can be derived from \code{E.info} and \code{dat}.}

\item{nF0}{TODO}

\item{verbose}{whether to print running messages}

\item{allow_empty}{whether to allow empty variable set}
}
\description{
Select cross-tissue via SODA
}
