% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_weights.R
\name{soda_model_imp_single}
\alias{soda_model_imp_single}
\title{Select tissue-specific via SODA}
\usage{
soda_model_imp_single(j, fix, E, dat, pos, gam = 1, scale = F, nF0 = 5)
}
\arguments{
\item{j}{the index of gene}

\item{fix}{a list of set containing the fixed effect for the gene}

\item{E}{here E is just for specific cell type, it is a matrix instead of a list}

\item{dat}{a list with three named elements. \describe{
\item{bed}{a matrix of size n_sample x n_snp}
\item{bim}{a data frame with n_snp rows, each row is the info for the snp}}}

\item{pos}{a list of genes, each element is a vector containing the SNP index that
locate in the gene, and it corresponds to each row of \code{E.info}.
It can be derived from \code{E.info} and \code{dat}.}

\item{gam}{tuning parameter \eqn{\gamma} in EBIC criteria. Default is 1.}

\item{scale}{whether to scale the expression and the genetic data when regressing the expression on the genetic data. Default is \code{FALSE}.}

\item{nF0}{minimum size of the set}
}
\description{
Select tissue-specific via SODA
}
