% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twas.R
\name{twas.multi.trait}
\alias{twas.multi.trait}
\title{Run association test for multiple traits}
\usage{
twas.multi.trait(list.stats, x, list.ts.eQTL, ncores = 1)
}
\arguments{
\item{list.stats}{a list of summary statistics, or a list of files to the summary statistics}

\item{x}{twas data list}

\item{list.ts.eQTL}{a list of tissue-specific eQTL for each tissue}

\item{ncores}{number of cores in parallel computing.}
}
\value{
a list of association tests for all traits
}
\description{
Run association test for multiple traits
}
