% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twas.R
\name{twas.res}
\alias{twas.res}
\title{MTWAS main function}
\usage{
twas.res(stats, soda.all.imp, dat, pos)
}
\arguments{
\item{stats}{summary statistics}

\item{soda.all.imp}{a list of all selected important genes}

\item{dat}{description}

\item{pos}{xx}
}
\description{
MTWAS main function
}
