% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twas.R
\name{twas.single.trait}
\alias{twas.single.trait}
\title{An association test in a tissue-specific manner for a phenotype (trait).}
\usage{
twas.single.trait(stats, x, list.ts.eQTL)
}
\arguments{
\item{stats}{summary statistics for a phenotype. It can be a filename to be read or a data.frame.
Suppose it has 5 columns with colnames \code{rsid}, \code{a1}, \code{a2}, \code{chr}, \code{z}.}

\item{x}{data}

\item{list.ts.eQTL}{a list of \code{ts.eQTL}, each element corresponds to a tissue}
}
\value{
z-statistics and p-values. Both are of size n_gene x n_tissue.
}
\description{
An association test in a tissue-specific manner for a phenotype (trait).
}
