% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twas_data}
\alias{twas_data}
\title{Datasets
Demo TWAS data of three genes and two tissues}
\format{
## `twas_data`
}
\usage{
twas_data
}
\description{
Datasets
Demo TWAS data of three genes and two tissues
}
\keyword{datasets}
